+++ src/Extension/PaginationExtension.php
@@ -126,7 +126,11 @@ final class PaginationExtension implements QueryResultCollectionExtension
         $metadata = $this->managerRegistry->getManagerForClass($operationClassName)?->getClassMetadata();
         if (!is_null($metadata)) {
             $rootAlias = $selectBuilder->getRootAlias();
-            foreach ($selectBuilder->getJoins()[$rootAlias] as $join) {
+            $joins = $selectBuilder->getJoins();
+            if(empty($joins)) {
+                return false;
+            }
+            foreach ($joins[$rootAlias] as $join) {
                 $joinData = explode('.', $join->join);
                 $mapping = $metadata->getAssociationMapping($joinData[1]);
                 if ($mapping['type'] === AssociationType::TO_MANY) {